+++ web/demo/App.tsx
@@ -1,6 +1,6 @@
 import type { FC } from 'react'
 
-import { CodyWebPanel, CodyWebPanelProvider, type Repository } from '../lib'
+import { CodyWebPanel, CodyWebPanelProvider, type InitialContext } from '../lib'
 
 // @ts-ignore
 import AgentWorker from '../lib/agent/agent.worker.ts?worker'
@@ -22,25 +22,13 @@ const serverEndpoint = localStorage.getItem('serverEndpoint') || DOTCOM_SERVER_E
 const accessTokenStorageKey = `accessToken:${serverEndpoint}`
 let accessToken = localStorage.getItem(accessTokenStorageKey)
 
-// Only for testing/demo purpose, in real-life usage consumer
-// should provide context repo information for Cody chat component
-const MOCK_DOT_COM_SOURCEGRAPH_REPOSITORY: Repository[] =
-    serverEndpoint === DOTCOM_SERVER_ENDPOINT
-        ? [
-              {
-                  id: 'UmVwb3NpdG9yeTozNjgwOTI1MA==',
-                  name: 'github.com/sourcegraph/sourcegraph',
-              },
-          ]
-        : []
-
-const MOCK_INITIAL_DOT_COM_CONTEXT =
-    serverEndpoint === DOTCOM_SERVER_ENDPOINT
-        ? {
-              fileURL: 'internal/codeintel/ranking/internal/background/mapper/config.go',
-              repositories: MOCK_DOT_COM_SOURCEGRAPH_REPOSITORY,
-          }
-        : undefined
+const MOCK_INITIAL_DOT_COM_CONTEXT: InitialContext = {
+    fileURL: 'web/demo/App.tsx',
+    repository: {
+        id: 'UmVwb3NpdG9yeTo2MTMyNTMyOA==',
+        name: 'github.com/sourcegraph/cody',
+    },
+}
 
 if (!accessToken) {
     accessToken = window.prompt(`Enter an access token for ${serverEndpoint}:`)