+++ web/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.6.0
+- Changes API for initial context (now it supports only one repository as an initial context mention)
+- Fixes problem with file and symbol mention search don't respect initial repository as a context
+
 ## 0.5.2 
 - Improves repository mention result ordering (for repositories, files and directories mentions)